requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
pycryptodome>=3.18.0
tkinterdnd2>=0.3.0
cryptography>=41.0.0
